Add algebra checks for generated kernels

+++ include/Canvas/Core/Specs.hpp
@@ -15,9 +15,9 @@ namespace canvas {
 
 /// Specifications for a single kernel.
 struct KernelSpecs {
-    size_t c, h, w;
+    int c, h, w;
 
-    KernelSpecs(size_t c, size_t h, size_t w): c(c), h(h), w(w) {}
+    KernelSpecs(int c, int h, int w): c(c), h(h), w(w) {}
 
     [[nodiscard]] size_t Hash() const {
         size_t hash = c;
@@ -35,7 +35,7 @@ private:
 
 public:
     size_t c_gcd = 0;
-    std::vector<size_t> c_gcd_factors;
+    std::vector<int> c_gcd_factors;
     std::vector<KernelSpecs> kernel_specs;
 
     explicit NetSpecs(std::vector<KernelSpecs> kernel_specs);
